add email sender service, refactor code, delete wildcard imports

+++ src/main/java/com/javadevs/testingservice/service/StudentService.java
@@ -4,9 +4,9 @@ import com.javadevs.testingservice.exception.StudentNotFoundException;
 import com.javadevs.testingservice.exception.SubjectNotFoundException;
 import com.javadevs.testingservice.model.Student;
 import com.javadevs.testingservice.model.Subject;
+import com.javadevs.testingservice.model.command.create.CreateStudentCommand;
 import com.javadevs.testingservice.model.command.edit.EditStudentCommand;
 import com.javadevs.testingservice.model.command.studentEdit.AddSubjectCoveredToStudentCommand;
-import com.javadevs.testingservice.model.command.create.CreateStudentCommand;
 import com.javadevs.testingservice.model.command.studentEdit.DeleteSubjectCoveredFromStudentCommand;
 import com.javadevs.testingservice.repository.QuestionRepository;
 import com.javadevs.testingservice.repository.StudentRepository;
@@ -56,7 +56,7 @@ public class StudentService {
     }
 
     @Transactional
-    public Student editStudentPartially(int id, EditStudentCommand command) {
+    public Student editStudentPartially(long id, EditStudentCommand command) {
         Student student = studentRepository.findById(id)
                 .map(studentToEdit -> {
                     ofNullable(command.getName()).ifPresent(studentToEdit::setName);
